updated with onlyoffice hashing
