+++ backend/swagger/docs/swagger.json
@@ -1178,6 +1178,7 @@
                     "type": "string"
                 },
                 "path": {
+                    "description": "index path",
                     "type": "string"
                 },
                 "size": {
