+++ backend/swagger/docs/swagger.yaml
@@ -14,6 +14,7 @@ definitions:
       name:
         type: string
       path:
+        description: index path
         type: string
       size:
         type: integer
